+++ tisha-wesbite/src/Components/Navigation/NavBar.js
@@ -10,12 +10,13 @@ import Container from '@mui/material/Container';
 import Button from '@mui/material/Button';
 import MenuItem from '@mui/material/MenuItem';
 import { common } from '@mui/material/colors';
-import BooksButton from '../BooksDisplay/BooksButton';
+
 import SubscribeButton from '../SubscribeButtonDisplay/SubscribeButton';
 import PropTypes from 'prop-types';
 import CssBaseline from '@mui/material/CssBaseline';
 import useScrollTrigger from '@mui/material/useScrollTrigger';
 import Slide from '@mui/material/Slide';
+import BooksButton from '../Books/BooksButton';
 
 function HideOnScroll(props) {
 	const { children, window } = props;
